Here I have created the header of my website, I have divided all the content in "div" elements with apropriate class names, in order to bea able to reference them when I need to style the content or animate it.
All my social media links will be from "font aswome" as their easier to 
insert and it gives a homogeneous look trough out the website.

I will be also inserting sections with Ids as I will create 4 diferent 
pages on the same page, and this way it is possible to tell the browser 
that the sections are different and associate it with the header 
sections.